+++ doc/api/search/all_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['abstractmockdbpersistence',['AbstractMockDbPersistence',['../class_pip_services_1_1_data_1_1_persistence_1_1_abstract_mock_db_persistence.html',1,'PipServices::Data::Persistence']]]
+  ['abstractmockdbpersistence',['AbstractMockDbPersistence',['../class_pip_services_1_1_data_1_1_persistence_1_1_abstract_mock_db_persistence.html',1,'PipServices3::Data::Persistence']]]
 ];