+++ doc/api/search/all_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['loadasync',['LoadAsync',['../interface_pip_services_1_1_data_1_1_i_loader.html#a0a9ebdf4459098c11c4314c2be12c566',1,'PipServices.Data.ILoader.LoadAsync()'],['../class_pip_services_1_1_data_1_1_persistence_1_1_json_file_persister.html#a9dac762677838a8c8c064a5a0a79be03',1,'PipServices.Data.Persistence.JsonFilePersister.LoadAsync()']]]
+  ['loadasync',['LoadAsync',['../interface_pip_services_1_1_data_1_1_i_loader.html#a0a9ebdf4459098c11c4314c2be12c566',1,'PipServices3.Data.ILoader.LoadAsync()'],['../class_pip_services_1_1_data_1_1_persistence_1_1_json_file_persister.html#a9dac762677838a8c8c064a5a0a79be03',1,'PipServices3.Data.Persistence.JsonFilePersister.LoadAsync()']]]
 ];