+++ doc/api/search/classes_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['jsonfilepersister',['JsonFilePersister',['../class_pip_services_1_1_data_1_1_persistence_1_1_json_file_persister.html',1,'PipServices::Data::Persistence']]]
+  ['jsonfilepersister',['JsonFilePersister',['../class_pip_services_1_1_data_1_1_persistence_1_1_json_file_persister.html',1,'PipServices3::Data::Persistence']]]
 ];