+++ doc/api/search/functions_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['filepersistence',['FilePersistence',['../class_pip_services_1_1_data_1_1_persistence_1_1_file_persistence.html#acf01964d561bcbb5f1795ce1dc066240',1,'PipServices.Data.Persistence.FilePersistence.FilePersistence(JsonFilePersister&lt; T &gt; persister)'],['../class_pip_services_1_1_data_1_1_persistence_1_1_file_persistence.html#a435cab710dfe691c2718333446a6a8fc',1,'PipServices.Data.Persistence.FilePersistence.FilePersistence()']]]
+  ['filepersistence',['FilePersistence',['../class_pip_services_1_1_data_1_1_persistence_1_1_file_persistence.html#acf01964d561bcbb5f1795ce1dc066240',1,'PipServices3.Data.Persistence.FilePersistence.FilePersistence(JsonFilePersister&lt; T &gt; persister)'],['../class_pip_services_1_1_data_1_1_persistence_1_1_file_persistence.html#a435cab710dfe691c2718333446a6a8fc',1,'PipServices3.Data.Persistence.FilePersistence.FilePersistence()']]]
 ];