+++ doc/api/search/properties_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['path',['Path',['../class_pip_services_1_1_data_1_1_persistence_1_1_json_file_persister.html#a27916ffe5ccbda6dc7e1dae9addc1bbb',1,'PipServices::Data::Persistence::JsonFilePersister']]]
+  ['path',['Path',['../class_pip_services_1_1_data_1_1_persistence_1_1_json_file_persister.html#a27916ffe5ccbda6dc7e1dae9addc1bbb',1,'PipServices3::Data::Persistence::JsonFilePersister']]]
 ];