+++ src/IPartialUpdater.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
-using PipServices.Commons.Data;
+using PipServices3.Commons.Data;
 
-namespace PipServices.Data
+namespace PipServices3.Data
 {
     /// <summary>
     /// Interface for data processing components to update data items partially.