+++ src/IQuerableReader.cs
@@ -1,8 +1,8 @@
-﻿using PipServices.Commons.Data;
+﻿using PipServices3.Commons.Data;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace PipServices.Data
+namespace PipServices3.Data
 {
     /// <summary>
     /// Interface for data processing components that can query a list of data items.