+++ src/Persistence/FilePersistence.cs
@@ -1,6 +1,6 @@
-﻿using PipServices.Commons.Config;
+﻿using PipServices3.Commons.Config;
 
-namespace PipServices.Data.Persistence
+namespace PipServices3.Data.Persistence
 {
     /// <summary>
     /// Abstract persistence component that stores data in flat files
@@ -17,7 +17,7 @@ namespace PipServices.Data.Persistence
     /// 
     /// ### References ###
     /// 
-    /// - *:logger:*:*:1.0   (optional) <a href="https://rawgit.com/pip-services-dotnet/pip-services-components-dotnet/master/doc/api/interface_pip_services_1_1_components_1_1_log_1_1_i_logger.html">ILogger</a> components to pass log messages
+    /// - *:logger:*:*:1.0   (optional) <a href="https://rawgit.com/pip-services3-dotnet/pip-services3-components-dotnet/master/doc/api/interface_pip_services_1_1_components_1_1_log_1_1_i_logger.html">ILogger</a> components to pass log messages
     /// </summary>
     /// <typeparam name="T">the class type</typeparam>
     /// <example>