+++ test/Persistence/IdentifiableFilePersistenceTest.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Threading;
-using PipServices.Commons.Config;
+using PipServices3.Commons.Config;
 using Xunit;
 
-namespace PipServices.Data.Persistence
+namespace PipServices3.Data.Persistence
 {
     public sealed class IdentifiableFilePersistenceTest : IDisposable
     {